mysql: update after review, MariaDB GTID filtering for binlog-push
